+++ app.trading.algoritmico.web/src/environments/environment.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: true,
   apiUrl: '',
-  version: '0.14.0'
+  version: '0.15.0'
 };